changes from review, thanks @agourlay
